WebSockets/tests/submissions/Opera/opening-handshake/003.html

<!doctype html>
<title>WebSockets: origin</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../constants.js></script>
<div id=log></div>
<script>
test(function(){
  var ws = new WebSocket(SCHEME_AND_DOMAIN+':8007/origin');
  ws.onmessage = function(e) {
    assert_equals(e.data, location.protocol+'//'+location.host);
    ws.onclose = function(e) {
      debug(e);
      assert_equals(e.wasClean, true);
      ws.onclose = assertUnreached;
      setTestTimeout(null);
      setTimeout(end, 50);
    }
    ws.close();
  }
  ws.onerror = ws.onclose = assertUnreached;
});
</script>
